Changed python files so that they allow for definition of libgomp and libstdc++ to be found locally and propagated into the python call.
CMakeLists.txt now generate this information and inserts it into the installed files.

Moved the Kokkos::initialize() and Kokkos::finalize() calls out of the Python_GMLS
class because multiple instantiations of the class in one test would cause an
error with instantiating and finalizing then reinstantiating cuda.

Now there is a wrapper so that python scripts can call initializeKokkos() and
finalizeKokkos() once at the beginning and end of the python script.