Changed "manifold [porder, weighting power, weighting type]" to "curvature [porder, weighting power, weighting type]" as this is more explicitly related to the calculation of the curvature, not the overall manifold problem that is being solved.
Removed "gmls" as a group in the XML as this leads to too much
confusion between the "remap" and "gmls" groups having the same
parameters. Now, everything is in "remap"

Renamed CompadreTrilinosTest.cpp to Compadre_TrilinosTest.cpp as well
as CompadreGMLSTest.cpp to Compadre_GMLSTest.cpp to match the other
tests.

Added Doxygen documentation to many of the private member functions in
the GMLS class.