Added applyAlphasToDataSingleComponentSingleTargetSite, which takes in a 1D View for input and calculates the target value at a single output site for a single input/output component corresponding to the target functional.
Added applyAlphasToDataSingleComponentAllTargetSites, which takes in a 1D View
for input and output (same memory space, but can be host or device) and calculates
the target value at all target sites, but again for a single input/output
component corresponding to the target functional.

Added applyAlphasToDataAllComponentsAllTargetSites, which effectively handles
all information that remap usually does, i.e. looping over input and output
components of the target operator and performing the necessary contractions
on data and repeatedly calling applyAlphasToDataSingleComponentAllTargetSites.