Changed _T matrix in GMLS class to a rank 3 tensor, with first rank being the target index, the second being the local tangent direction, and the third being the spatial dimension. This has the effect of transposing T as it was previously being used.
Added the ability to set the tangent directions with a Kokkos View
rather than having to calculate them with an approximation from
the nearest neighbors in the point cloud. Default use of the toolkit
on manifolds will continue to calculate the local tangent directions.

Added an example to the GMLS_Manifold_Test that performs the
reconstruction of a vector manifold. The data is sampled in ambient
space, projected to the local space, reconstructed in the local space
using a vector basis in GMLS, and then projected back to ambient
space.