Added Enum describing the effective and actual rank of a polynomial basis. This allows for scalar solutions to be reused for a vector basis, while using consistent dimension arguments.
Added a test to GMLS_Staggered_Manifold to check staggered divergence
using both a scalar and vector, with two different sampling functionals.

Renamed ManifoldVectorPointSample to VectorPointSample and made it
consistent across both manifold and nonmanifold problems.

Added a space named VectorOfScalarClonesTaylorPolynomial which solves
for a scalar solution that is then reused with different target
evaluations dependent on component to produce the same result as a
full vector basis solve, with a substantially smaller cost.
- On 6th order, about 45% less work. On 9th order, 20x faster.