Added optional LAPACK_LOCATION for when a specific version of the LAPACK library is needed, but doesn't end in .so.
Added optional LAPACK_DECLARED_THREADSAFE in order to allow a user to
declare that their version of LAPACK is not threadsafe. This means that
a LAPACK call cannot be wrapped in a parallel for. Ideally, all LAPACK
libraries would be built with -frecursive, but that is not the case.

Running with LAPACK_DECLARED_THREADSAFE:BOOL=OFF will signifcantly slow
down computations for users without a threadsafe LAPACK.