Added a static member function to PointCloudSearch that allows user to get a good upper bound estimate which is convenient for preallocating the neighbor_lists Kokkos view.
Switched to using this in the examples prior to performing neighbor
search.

Added a RadiusResultSet class overload for nanoflann that allows us to
specify a preallocated vector of doubles and ints along with their
maximum size, and fill these in with neighbor search. This replaces the
nanoflann default which uses std::vector and allocates from the heap and
uses push to add neighbors.