Added gradient target operator to lists of target operators registered in Python interface.
Added an extra string argument in applyStencil function to choose the
x derivative of the gradient as the target to be applied.

Added a resetCoefficientData function to the GMLS class that is called
any time a modifier member function of the class is called.
- This allows the GMLS class instance to be reused
- Ensures that the polynomial coefficients for a reconstruction
  are destroyed once the data used to generate the coefficients has
  changed.

Made a GMLS_SmallBatchReuse_Device_Test which exercises repeatedly
reusing a single instance of a GMLS class, changing the target site,
the window size, and the single neighbor list.