Added GMLS_Vector test to exercise the VectorTaylorPolynomial basis.
Refactored the GMLS_Targets (not on manifolds) to be organized by
reconstruction space rather than by operation.

Added a setExtraData function allowing a rank 2 matrix to be
available to sampling functionals acting on a basis and on target
operations.

Added SamplingFunctional::FaceNormalIntegralSample and
SamplingFunctional::FaceNormalIntegralSampleData.

Removed the enum EffectiveReconstructionSpaceRank as it was never used.