Removed pyproject.toml file to remove compliance with PEP 517 and 518 standards for package installation (which also implies an isolated build).
It is now possible that there will be unresolved dependencies for the
Python module build for Compadre, but these will be caught by the setup.py
which will return a message to the user indicating missing dependencies
and how to quickly satisfy them.

This effectively adds a --no-build-isolation flag to pip, which allows
the use of existing wheels.