Created OptimizationAlgorithm enum with options {NONE, OBFET, CAAS}
Created OptimizationObject class which allows for specifying an
OptimizationAlgorithm and a global lower and upper optional bound.

Converted OBFET class to OptimizationManager.
Now uses OptimizationObjects instead of a bool to turn on optimization.
Modified RemoteRemapManager to pass all pieces needed to form a
OptimizationObject.

Switched a test using OBFET to turn on for a field that has its accuracy
measured as part of the test.