Made changes to indexing to make compatible with Kokkos in Trilinos, as some deprecated functionality had been removed which we were using.
Turned off "preserve gids" in VTK type file reading since it doesn't
work as intended. Future issue to fix this and add a test.

Changed TrilinosTest for the write/reread flag check to now use the
gid as the flag and then check that it matches when read back in.