Modified tribits tests setting NUM_MPI_PROCS to 1 for each and enabling both mpi and serial for COMM.
Changed python based tests to use PYTHON_EXECUTABLE and to be created
as an advanced tests which allows for avoiding the call to python by
mpiexec.

Compadre_Config.h is now produced using bob features for both raw
CMake and Tribits build. Modified bob call to allow for a module name
rather than it being hard coded for ${PROJECT_NAME}.