Added keep_coefficients argument to generateAlphas() and generatePolynomialCoefficients() which is set to false by default. If false, then solution to (PTWP^-1 * PTW deallocated after its use in generating the alpha coefficients.
Descriptive error is given when trying to access coefficients that are
chosen not to be stored. Check added that if storing coefficients,
number of batches must be equal to 1 (otherwise all coefficients weren't
stored, anyways).

Tests that access polynomial coefficients have been modified to use the
true flag for keep_coefficients when calling generateAlphas().