Added NeighborListAccessor class which will take a compressed row neighborhood lists and allow it to be accessible in the same way as the 2D array neighborhood lists.
Added a conversion function which will convert a 2D neighborhood
lists to a compressed row storage neighborhood list.

Added tests to the NeighborSearchTest to verify that the compressed
row and 2D neighborhood lists both work as well as a 2D array
neighborhood list that is then converted to a compressed row format.