Bumped version to 1.2.0.
Changed the neighbor search in pycompadre to use the compressed row
storage format and also used the dry-run feature to determine the
storage size requirements rather than using some storage_multiplier
as was done before. This is now much more robust for neighbor search
in pycompadre.

Wrapped the NeighborLists class in pycompadre so that it could be
returned by ParticleHelper::getNeighbors which has now been renamed
ParticleHelper::getNeighborLists.