Change interface all physics, sources, and boundary conditions
Updated Newmark scheme to enable use of a Dirichlet displacement
condition to set acceleration values at the previous timestep.

In addition to simulation time, which previously was provided, now
the current timestep size as well as the previous timestep size are
provided to all physics, sources, and boundary conditions.

This should be useful in the future for variable size timesteps and
will work for both RK and Newmark schemes.